TB functional tests: Merge tbselenium dir
