+++ securedrop/static/i/custom_logo.png
