--- securedrop/tbselenium/functional/test_source_notfound.py
@@ -1,10 +1,0 @@
-import source_navigation_steps
-import functional_test
-
-
-class TestSourceInterfaceNotFound(
-        functional_test.FunctionalTest,
-        source_navigation_steps.SourceNavigationStepsMixin):
-
-    def test_not_found(self):
-        self._source_not_found()
