+++ securedrop/tests/functional/functional_test.py
@@ -1,23 +1,35 @@
 # -*- coding: utf-8 -*-
 
+from datetime import datetime
 import errno
-import mock
+from multiprocessing import Process
 import os
+from os.path import abspath, dirname, join, realpath, expanduser
 import signal
 import socket
 import time
+import json
 import traceback
+import subprocess
+import shutil
 import requests
 
-from datetime import datetime
-from multiprocessing import Process
-from os.path import abspath, dirname, join, realpath
+
+
+import pyotp
+import gnupg
 from selenium import webdriver
 from selenium.common.exceptions import (WebDriverException,
                                         NoAlertPresentException)
-from selenium.webdriver.firefox import firefox_binary
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions
+
+TBB_PATH = abspath(join(expanduser('~'), '.local/tbb/tor-browser_en-US/'))
+os.environ['TBB_PATH'] = TBB_PATH
+TBBRC = join(TBB_PATH, 'Browser/TorBrowser/Data/Tor/torrc')
+
+from tbselenium.tbdriver import TorBrowserDriver
+from tbselenium.utils import start_xvfb, stop_xvfb
 
 os.environ['SECUREDROP_ENV'] = 'test'  # noqa
 from sdconfig import config
@@ -25,9 +37,11 @@ import journalist_app
 import source_app
 import tests.utils.env as env
 
-from db import db
 
-LOG_DIR = abspath(join(dirname(realpath(__file__)), '..', 'log'))
+FUNCTIONAL_TEST_DIR = abspath(dirname(__file__))
+LOGFILE_PATH = abspath(join(FUNCTIONAL_TEST_DIR, 'log/firefox.log'))
+FILES_DIR = abspath(join(dirname(realpath(__file__)), '../..', 'tests/files'))
+
 
 
 # https://stackoverflow.com/a/34795883/837471
@@ -46,21 +60,55 @@ class FunctionalTest(object):
 
     def _unused_port(self):
         s = socket.socket()
-        s.bind(("localhost", 0))
+        s.bind(("127.0.0.1", 0))
         port = s.getsockname()[1]
         s.close()
         return port
 
-    def _create_webdriver(self, firefox, profile=None):
+    def add_hidservauth(self, address, token):
+        if not os.path.exists(TBBRC):
+            return False
+        found_flag = False
+        entry = "HidServAuth {0} {1}\n".format(address,token)
+        lines = []
+        with open(TBBRC) as fobj:
+            lines = fobj.readlines()
+        for line in lines:
+            if entry.strip() == line.strip():
+                found_flag = True
+
+        if found_flag:  # We already have the information in the torrc file
+            return True
+        lines.append(entry)
+
+        with open(TBBRC, 'w') as fobj:
+            fobj.write(''.join(lines))
+        return True
+
+    def _create_webdriver(self,  profile=None):
+        log_file = open(LOGFILE_PATH, 'a')
+        log_file.write(
+            '\n\n[%s] Running Functional Tests\n' % str(
+                datetime.now()))
+        log_file.flush()
         # see https://review.openstack.org/#/c/375258/ and the
         # associated issues for background on why this is necessary
         connrefused_retry_count = 3
         connrefused_retry_interval = 5
 
+        # Don't use Tor when reading from localhost, and turn off private
+        # browsing. We need to turn off private browsing because we won't be
+        # able to access the browser's cookies in private browsing mode. Since
+        # we use session cookies in SD anyway (in private browsing mode all
+        # cookies are set as session cookies), this should not affect session
+        # lifetime.
+        pref_dict = {'network.proxy.no_proxies_on': '127.0.0.1',
+                        'browser.privatebrowsing.autostart': False}
         for i in range(connrefused_retry_count + 1):
             try:
-                driver = webdriver.Firefox(firefox_binary=firefox,
-                                           firefox_profile=profile)
+                driver = TorBrowserDriver(TBB_PATH,
+                                        pref_dict = pref_dict,
+                                        tbb_logfile_path = LOGFILE_PATH)
                 if i > 0:
                     # i==0 is normal behavior without connection refused.
                     print('NOTE: Retried {} time(s) due to '
@@ -73,69 +121,81 @@ class FunctionalTest(object):
                     continue
                 raise
 
-    def _prepare_webdriver(self):
-        log_file = open(join(LOG_DIR, 'firefox.log'), 'a')
-        log_file.write(
-            '\n\n[%s] Running Functional Tests\n' % str(
-                datetime.now()))
-        log_file.flush()
-        return firefox_binary.FirefoxBinary(log_file=log_file)
+    def _create_secondary_firefox_driver(self):
+        profile = None
+        self.f_profile_path = "/tmp/testprofile"
+        if os.path.exists(self.f_profile_path):
+            shutil.rmtree(self.f_profile_path)
+        if self.journalist_location.find('.onion') != -1:
+            os.mkdir(self.f_profile_path)
+            profile = webdriver.FirefoxProfile(self.f_profile_path)
+            # set FF preference to socks proxy in Tor Browser
+            profile.set_preference("network.proxy.type", 1)
+            profile.set_preference("network.proxy.socks", "127.0.0.1")
+            profile.set_preference("network.proxy.socks_port", 9150)
+            profile.set_preference("network.proxy.socks_version", 5)
+            profile.set_preference("network.proxy.socks_remote_dns", True)
+            profile.set_preference("network.dns.blockDotOnion", False)
+            profile.update_preferences()
+        self.second_driver = webdriver.Firefox(firefox_profile=profile)
+        self.second_driver.implicitly_wait(15)
+
+    def swap_drivers(self):
+        if not self.second_driver:
+            self._create_secondary_firefox_driver()
+        # Only if we two drivers
+        if self.driver and self.second_driver:
+            self.driver, self.second_driver = self.second_driver, self.driver
+
+    def init_gpg(self):
+        """Initialize the GPG keyring and import the journalist key for
+        testing.
+        """
+        gpg = gnupg.GPG(homedir="/tmp/testgpg")
+        # Faster to import a pre-generated key than to gen a new one every time.
+        for keyfile in (join(FILES_DIR, "test_journalist_key.pub"),
+                        join(FILES_DIR, "test_journalist_key.sec")):
+            gpg.import_keys(open(keyfile).read())
+        return gpg
+
+    def system(self, cmd):
+        """
+        Invoke a shell command. Primary replacement for os.system calls.
+        """
+        print(cmd)
+        ret = subprocess.Popen(cmd, shell=True, stdin=subprocess.PIPE,
+                               stdout=subprocess.PIPE, stderr=subprocess.PIPE, close_fds=True)
+        out, err = ret.communicate()
+        return out
 
     def setup(self, session_expiration=30):
-        self.__context = journalist_app.create_app(config).app_context()
-        self.__context.push()
-        # Patch the two-factor verification to avoid intermittent errors
-        self.patcher = mock.patch('models.Journalist.verify_token')
-        self.mock_journalist_verify_token = self.patcher.start()
-        self.mock_journalist_verify_token.return_value = True
-
-        self.patcher2 = mock.patch('source_app.main.get_entropy_estimate')
-        self.mock_get_entropy_estimate = self.patcher2.start()
-        self.mock_get_entropy_estimate.return_value = 8192
 
         signal.signal(signal.SIGUSR1, lambda _, s: traceback.print_stack(s))
 
-        env.create_directories()
-        self.gpg = env.init_gpg()
-        db.create_all()
 
-        source_port = self._unused_port()
-        journalist_port = self._unused_port()
+        instance_information_path = join(FUNCTIONAL_TEST_DIR, 'instance_information.json')
 
-        self.source_location = "http://localhost:%d" % source_port
-        self.journalist_location = "http://localhost:%d" % journalist_port
+        if os.path.exists(instance_information_path):
+            with open(instance_information_path) as fobj:
+                data = json.load(fobj)
+            self.source_location = data.get('source_location', "http://127.0.0.1:8080")
+            self.journalist_location = data.get('journalist_location', "http://127.0.0.1:8081")
+            self.hidservauth = data.get('hidserv_token', '')
+            self.admin_user = data.get('user')
+            self.admin_user['totp'] = pyotp.TOTP(self.admin_user['secret'])
+            self.new_totp = None # To be created runtime
+            self.sleep_time = data.get('sleep_time', 10)
+            if self.hidservauth:
+                if self.journalist_location.startswith('http://'):
+                    location = self.journalist_location[7:]
+                self.add_hidservauth(location, self.hidservauth)
+        else:
+            assert False, "Missing instance information JSON file."
 
         # Allow custom session expiration lengths
         self.session_expiration = session_expiration
 
-        self.source_app = source_app.create_app(config)
-        self.journalist_app = journalist_app.create_app(config)
-
-        def start_source_server(app):
-            config.SESSION_EXPIRATION_MINUTES = self.session_expiration
-
-            app.run(
-                port=source_port,
-                debug=True,
-                use_reloader=False,
-                threaded=True)
-
-        def start_journalist_server(app):
-            app.run(
-                port=journalist_port,
-                debug=True,
-                use_reloader=False,
-                threaded=True)
-
-        self.source_process = Process(
-            target=lambda: start_source_server(self.source_app))
-
-        self.journalist_process = Process(
-            target=lambda: start_journalist_server(self.journalist_app))
-
-        self.source_process.start()
-        self.journalist_process.start()
-
+        """
         for tick in range(30):
             try:
                 requests.get(self.source_location)
@@ -144,9 +204,12 @@ class FunctionalTest(object):
                 time.sleep(1)
             else:
                 break
+        """
+        self._create_secondary_firefox_driver()
+        #  if not hasattr(self, 'override_driver'):
+        self.driver = self._create_webdriver()
 
-        if not hasattr(self, 'override_driver'):
-            self.driver = self._create_webdriver(self._prepare_webdriver())
+        self.gpg = self.init_gpg()
 
         # Polls the DOM to wait for elements. To read more about why
         # this is necessary:
@@ -161,32 +224,27 @@ class FunctionalTest(object):
         # will only report failure after 60 seconds which is painful
         # for quickly debuging.
         #
-        self.driver.implicitly_wait(10)
+        self.driver.implicitly_wait(15)
+
 
         # Set window size and position explicitly to avoid potential bugs due
         # to discrepancies between environments.
-        self.driver.set_window_position(0, 0)
-        self.driver.set_window_size(1024, 768)
+        #self.driver.set_window_position(0, 0)
+        #self.driver.set_window_size(1024, 768)
 
         self.secret_message = ('These documents outline a major government '
                                'invasion of privacy.')
 
-    def wait_for_source_key(self, source_name):
-        filesystem_id = self.source_app.crypto_util.hash_codename(source_name)
-
-        def key_available(filesystem_id):
-            assert self.source_app.crypto_util.getkey(filesystem_id)
-        self.wait_for(
-            lambda: key_available(filesystem_id), timeout=60)
-
     def teardown(self):
-        self.patcher.stop()
-        env.teardown()
-        if not hasattr(self, 'override_driver'):
+
+        if self.driver:
             self.driver.quit()
-        self.source_process.terminate()
-        self.journalist_process.terminate()
-        self.__context.pop()
+        if self.second_driver:
+            self.second_driver.quit()
+
+    def create_new_totp(self, secret):
+        self.new_totp = pyotp.TOTP(secret)
+
 
     def wait_for(self, function_with_assertion, timeout=5):
         """Polling wait for an arbitrary assertion."""
