+++ securedrop/tests/functional/test_submit_and_retrieve_file.py
@@ -14,11 +14,13 @@ class TestSubmitAndRetrieveFile(
         self._source_continues_to_submit_page()
         self._source_submits_a_file()
         self._source_logs_out()
+        self.swap_drivers()  # To use Firefox
         self._journalist_logs_in()
         self._journalist_checks_messages()
         self._journalist_stars_and_unstars_single_message()
         self._journalist_downloads_message()
         self._journalist_sends_reply_to_source()
+        self.swap_drivers()  # To use Tor
         self._source_visits_source_homepage()
         self._source_chooses_to_login()
         self._source_proceeds_to_login()
