+++ securedrop/tests/functional/test_submit_and_retrieve_message.py
@@ -14,6 +14,7 @@ class TestSubmitAndRetrieveMessage(
         self._source_continues_to_submit_page()
         self._source_submits_a_message()
         self._source_logs_out()
+        self.swap_drivers()  # To use Firefox
         self._journalist_logs_in()
         self._journalist_checks_messages()
         self._journalist_downloads_message()
